move normalized state to subdir
